Ev2 shims:
Checkpoint some work regarding specialized sorting of paths for the
Ev1 drive within the shims. This revision sets up the mechanics for
sorting the changes and passing them into the path_driver.

However: 27 tests failed with the new sorting. Falling back to the
normal sort mechanic returns to the single merge 105 failure.

* subversion/libsvn_delta/compat.c:
  (sort_deletes_first): new function to sort paths where the deletes
    are delivered before any further adds, in order to account for
    case-only renames.
  (get_sorted_paths): helper function to extract a proper sort-order
    from the CHANGES hash.
  (drive_changes): use the new get_sorted_paths() helper. because it
    doesn't quite work, resort the paths. use the new path_driver
    function from r1332523.


git-svn-id: http://svn.apache.org/repos/asf/subversion/trunk@1332562 13f79535-47bb-0310-9956-ffa450edef68