On a path towards placing HTTP-level items in the right place, add a DONE flag to the handler, along with a captured error from the server. This handler-level feature will replace simple_request_context.
We also introduce a new function to create a request (for a handler)
and wait for its completion. This also encapsulates the DONE flag,
unlike context_run_wait().

Note: the new run_one() creates the request unlike run_wait().
Examining all of the run_wait() calls, we see a request immediately
preceeds it. May as well bottle it all up.

* subversion/libsvn_ra_serf/ra_serf.h:
  (svn_ra_serf__server_error_t): move the typedef higher in the file
  (svn_ra_serf__handler_t): add DONE and SERVER_ERROR.
  (svn_ra_serf__context_run_one): new declaration

* subversion/libsvn_ra_serf/util.c:
  (svn_ra_serf__context_run_one): new implementation


git-svn-id: http://svn.apache.org/repos/asf/subversion/trunk@1334697 13f79535-47bb-0310-9956-ffa450edef68