Add an 'svn_rangelist_t' typedef for use with the existing 'svn_rangelist_*' APIs, instead of using 'apr_array_header_t' directly.
The immediate reason is to improve the debugging experience: I can make
GDB display 'svn_rangelist_t' values in a simple human-readable form, whereas
the best it can do with 'apr_array_header_t', not knowing the type of the
elements, is print something like 'array of 2 items'.

Beyond that, it makes sense from a coding abstraction point of view, filling
a gap in this hierarchy:

  svn_mergeinfo_catalog_t   is a collection of
  svn_mergeinfo_t           is a collection of
  svn_rangelist_t           is a collection of
  svn_merge_range_t

This type has public visibility.  Because it is just a typedef, the C API
will be backward-compatible and the ABI won't change.  Therefore all uses,
even in deprecated functions, are being changed.

Using such a typedef doesn't provide any additional static type-checking.

Unlike svn_mergeinfo_t and svn_mergeinfo_catalog_t, svn_rangelist_t is not a
pointer type.

* subversion/include/svn_mergeinfo.h
  (svn_rangelist_t): New typedef.
  (...everywhere...): Use it.

* subversion/include/private/svn_mergeinfo_private.h
  (...everywhere...): Use it.

* subversion/include/svn_client.h
  (svn_client_mergeinfo_get_merged): Update doc string to refer to this.

* subversion/bindings/javahl/native/CreateJ.cpp,
  subversion/bindings/javahl/native/CreateJ.h,
  subversion/bindings/javahl/native/SVNClient.cpp,
  subversion/bindings/swig/core.i,
  subversion/bindings/swig/include/svn_containers.swg,
  subversion/bindings/swig/python/libsvn_swig_py/swigutil_py.c
  (...everywhere...): Update the bindings.

  subversion/libsvn_client/diff.c,
  subversion/libsvn_client/merge.c,
  subversion/libsvn_client/mergeinfo.c,
  subversion/libsvn_client/mergeinfo.h,
  subversion/libsvn_client/ra.c,
  subversion/libsvn_fs_util/fs-util.c,
  subversion/libsvn_repos/load-fs-vtable.c,
  subversion/libsvn_repos/log.c,
  subversion/libsvn_repos/rev_hunt.c,
  subversion/libsvn_subr/deprecated.c,
  subversion/libsvn_subr/mergeinfo.c,
  subversion/svndumpfilter/main.c,
  subversion/svnrdump/load_editor.c,
  subversion/tests/libsvn_subr/mergeinfo-test.c
  (...everywhere...): Use it.


git-svn-id: http://svn.apache.org/repos/asf/subversion/trunk@1367113 13f79535-47bb-0310-9956-ffa450edef68