Fix check-swig-py and check-swig-rb targets on OS X.
OS X needs the DYLD_LIBRARY_PATH env variable set in order to find the libraries
to link against.  In Python's case we have setup_path.py which is supposed to
be a good place to do this.  However, it won't work since it sets the
environment variables after python is already running and OS X's dynamic linker
only reads the environment variables at startup.

* Makefile.in
  (check-swig-py, check-swig-rb): Build a DYLD_LIBRARY_PATH variable if needed.


git-svn-id: http://svn.apache.org/repos/asf/subversion/trunk@1397846 13f79535-47bb-0310-9956-ffa450edef68