In the move-update editor, obtain info about the tree-conflict earlier, before we start using wcroot and relpath pairs.
This backpedals a bit from my goal to make all of this file use wcroot
and relpath pairs. But there are too many APIs that would need to be
converted to wcroot+relpath, and doing so wouldn't gain us much.
Reading tree-conflict info outside the wc_db transaction should be safe
since this code assumes that a write-lock is held on the working copy.

* subversion/libsvn_wc/wc_db_update_move.c
  (get_tc_info): Drop unused dst_abspath parameter.
  (update_moved_away_conflict_victim_baton): New members operation,
   local_change, incoming_change, old_version, and new_version.
  (update_moved_away_conflict_victim): Don't call get_tc_info() here.
   Pull tree-conflict info from the caller-provided baton instead.
  (svn_wc__db_update_moved_away_conflict_victim): Read tree-conflict info
   into update_moved_away_conflict_victim()'s baton, before starting the
   wc_db transaction.


git-svn-id: http://svn.apache.org/repos/asf/subversion/trunk@1403684 13f79535-47bb-0310-9956-ffa450edef68