Yet another server-side tool: it crawls through your format 4+ FSFS repository and prints data size & count stats.
This is the initial commit based on a copy of fsfs-reorg for
its FSFS structure reading and online representation code.
Lots of more info such as histograms, per file-type info etc
will be added in the future.

Not really sure where I want to go with that but for now, it
helps me verifying my assumptions on which data or structures
is dominant in FSFS repositories etc.  Since all access is
read-only, it should be safe to use but it still requires a
lot of memory.  Also not sure whether that's going to be a
long-term supported tool.

* build.conf
  (__ALL_TESTS__): add fsfs-stats dependency
  (fsfs-stats): new target / project section

* tools/server-side/fsfs-stats.c
  () new tool sources


git-svn-id: http://svn.apache.org/repos/asf/subversion/trunk@1410995 13f79535-47bb-0310-9956-ffa450edef68