Introduce SVN_SQLITE__WITH_LOCK macro.  A follow-up to r1423728 which introduced SVN_SQLITE__WITH_TXN and SVN_SQLITE__WITH_IMMEDIATE_TXN.
This provides a more convenient interface for wrapping a transaction than
the existing function svn_sqlite__with_lock(), because there is no need to
marshall the called function's parameters via a 'baton' structure.

Use it in the one place where it is both beneficial and in regularly
maintained code (that is, not in backward-compatibility code).  This will
serve as both a test and a showcase.  Further use -- through a new
SVN_WC__DB_WITH_TXN macro -- will follow in a separate commit.

* subversion/include/private/svn_sqlite.h
  (svn_sqlite__begin_savepoint,
   svn_sqlite__finish_savepoint): New functions.
  (SVN_SQLITE__WITH_LOCK): New macro.

* subversion/libsvn_subr/sqlite.c
  (svn_sqlite__begin_savepoint,
   svn_sqlite__finish_savepoint): New functions extracted from
    svn_sqlite__with_lock().
  (svn_sqlite__with_lock): Rewrite as a simple wrapper around the new macro.

* subversion/libsvn_wc/wc_db.c
  (init_db_baton): Delete.
  (init_db): Take the parameters directly.
  (create_db): Pass the parameters directly, using SVN_SQLITE__WITH_LOCK().


git-svn-id: http://svn.apache.org/repos/asf/subversion/trunk@1423944 13f79535-47bb-0310-9956-ffa450edef68