Following up on r855643 (from 2005-08-03), update the expected result of an fs test after the fix for issue #2751 "Directory prop mods reverted in overlapping commits scenario" (fixed in 1.5.0) and some later small fs api behavior changes.
The test expected a delete of an already deleted node to silently succeed and
also it expected svn_fs_commit_txn() to return a valid revision number even
when a commit fails, an assumption which was not always true and currently
never true.

Note that an earlier attempt to fix this test was applied in r875421
and later reverted, because it caused BDB problems. This patch works around
that problem by providing better base revisions like if a user has updated
his working copy, instead of by expecting more errors.

* subversion/tests/libsvn_fs/fs-test.c
  (merging_commit):
    5: Don't expect a valid revision number.
    4: Expect delete of deleted node to fail.
    (in deleted ancestor): Expect invalid revnum.
    Don't try to delete A/D/H a third time.

  (test_funcs): Expect merging_commit to pass.

git-svn-id: http://svn.apache.org/repos/asf/subversion/trunk@1426778 13f79535-47bb-0310-9956-ffa450edef68