Snapshot some testwork on issue #3362 by adding a second variant of a test and some disabled helper code.
Variant 2 currently shows different bugs in the file and memory
unified diff output. (The memory code misses the verify first line of the original
and the file code tries to read context from lines that have already been
processed, which the reader doesn't support. The result is a broken index in the
hunk header.

* subversion/tests/libsvn_diff/diff-diff3-test.c
  (two_way_diff): Add some disabled helper code.

  (test_issue_3362): Duplicate this test in..

  (two_way_issue_3362_v1): The original testset.

  (two_way_issue_3362_v2): The final variant.


git-svn-id: http://svn.apache.org/repos/asf/subversion/trunk@1437250 13f79535-47bb-0310-9956-ffa450edef68